docs(#23): WalkingMap 코드 정리
