+++ index.html
@@ -12,7 +12,7 @@
     <script type="module" src="/src/main.jsx"></script>
     <script
       type="text/javascript"
-      src="//dapi.kakao.com/v2/maps/sdk.js?appkey=356aa1c57243e551c6ea0db569b098cf"
+      src="//dapi.kakao.com/v2/maps/sdk.js?appkey=356aa1c57243e551c6ea0db569b098cf&libraries=services,geometry"
     ></script>
   </body>
 </html>
